Moved DeleteMonitorService to service package. Removed unnecessary double check when deleting the monitors (if the monitor belongs to another workflow)
